requirements.txt
fastapi==0.129.0
google-genai==1.65.0
python-dotenv==1.1.0
python-multipart==0.0.22
uvicorn==0.41.0
google-adk==1.25.1
httpx==0.28.1
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
asyncpg==0.30.0
PYJWT
supabase==2.22.0